Always use ar_archive_writer for import libs
